pyproject.toml
[tool.poetry]
name = "brainframe-cli"
version = "0.1.0.dev14"
license = "BSD-3-Clause"
authors = ["Aotu <info@aotu.ai>"]

description = "Makes installing and managing a BrainFrame server easy"
readme = "README.rst"

repository = "https://github.com/aotuai/brainframe_cli"
homepage = "https://github.com/aotuai/brainframe_cli"

packages = [
    { include = "brainframe/cli/**/*.py" },
]
include = ["brainframe/cli/translations/*"]

[tool.poetry.dependencies]
python = ">=3.6"
python-i18n = "^0.3"
pyyaml = "^5.3"
distro = "^1.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "*"

[tool.poetry.scripts]
brainframe = "brainframe.cli.main:main"

[tool.black]
line-length = 79
